Split model module into separate files for each model
